Finish omp_get_thread_num attempt with pthread_self
Additionally, determined that private vars are uninitialized upon
entry to the pragma region. This means that all that work on dynamically
creating structs was for nothing. There is no need to create a struct to
hold the values--simply redeclare them on the first line of the pthread
function.

Next commit will attempt to revert the dynamic struct code :(
The hash map will still be useful, as will most other functions.
Additionally, this knowledge about how to pass structs is still good.
For loops, you need to pass a struct with "start" and "end" vars so
it knows which iterations to do.