Move a few things
Adjust numthreads to save to a global so it's safe to call it
multiple times.

About to try a new strategy--rewrite the processing of parallel
and parallel for so that the input vector is not modified until
the end. This involves saving all new functions to strvecs and putting
those in a vector of strvecs inserted at the end of main.
The trick is that main will change due to pthread calls but that
might be OK (should come after adjustments to new functions)