Good commit!!!!!!!
This is a great fallback commit.
Regression testing confirms that par, parfor, and single RUN
on general. Only parfor has expected output and even that one
is off by one. par still has wrong threadnums and single is not
executing the single section for some reason. Step through
on linux may be the only way to know why. This is totally
doable!!

Additionally, commented out the code in the function which removes
var declarations from main. In single, those vars are used after
the parallel section and it appears there is no need to remove them.
Leaving it for now for easy restoration, and calls from
parallel and parfor are still there.