Fix off by one, verified
Fix off by one error in parfor and verified with regression
testing that this produces runnable files from par, parfor,
and single samples.

Single still appears to not work and outputs garbage, but note
that this commit fixes the issue with par! it outputs perfectly
now. The trick was not to create pthreads in a for loop, but to
explicitly create them as in parfor