Basically the same, but with a new problem
The problem is this: private variables are redeclared within
their functions but must remain declared in main because they
could be used later.

Shared variables are declared as global, which means they do not
need to remain declared in main and if they are, will output garbage.

There are two solutions to this:
1. Remove all declarations and re-add private ones. It is trivial
to remove all declarations but very difficult to know where to re-add
the private ones.
2. Do not remove any declarations initially, but go back later
and remove all global var declarations. This is also difficult, as
finding declarations is very tricky and the lines all change.

This commit was verified to function on general for par, parfor,
and single, but single prints garbage unless you manually remove
the redeclaration of global vars in main (a and b)